Doing more calculations in the worker

+++ server.mjs
@@ -8,8 +8,8 @@ console.log(cpus)
 const interval = []
 
 let index = 0
-while (index < 100000) {
-  interval.push({ start: index, end: index += 9999 })
+while (index < 1000000000) {
+  interval.push({ start: index, end: index += 99999999 })
   index++
 }
 
@@ -23,12 +23,14 @@ const NS_PER_SEC = 1e9
 const time = process.hrtime()
 
 worker.on('message', msg => {
-  console.log('msg: ', msg)
+  // console.log('msg: ', msg)
   const diff = process.hrtime(time)
-  console.log(`Benchmark took ${diff[0] + diff[1] / NS_PER_SEC} seconds`)
 
   console.log('il', interval.length)
-  interval.length || worker.terminate()
+  if(interval.length === 0) {
+    interval.length || worker.terminate()
+    console.log(`Benchmark took ${diff[0] + diff[1] / NS_PER_SEC} seconds`)
+  }
   interval.length && worker.postMessage(interval.pop())
 })
